+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-# Play Changelog
+# Changelog
+
+## 2.0.0 2016-08-15
+
+- Change resource_name default from servicename to source 
+- Default servicename to project_name when not defined 
+- Remove default recipe
+- Support systemd
+- Remove upstart script
+- Update systemv script
+- Remove ark cookbook dependency
 
 ## 1.0.2 2016-04-01